+++ lib/src/telegram/models/user.dart
@@ -1,8 +1,9 @@
 part of 'models.dart';
 
 /// This object represents a Telegram user or bot.
-class User {
+class User implements WithID {
   /// Unique identifier for this user or bot. This number may have more than 32 significant bits and some programming languages may have difficulty/silent defects in interpreting it. But it has at most 52 significant bits, so a 64-bit integer or double-precision float type are safe for storing this identifier.
+  @override
   final int id;
 
   /// True, if this user is a bot